+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - **button**: add `endIcon` and `startIcon` properties to allow adding an icon at the start or the end of the button text
 - **mat-icon**: add css variables for sizes of icons
+- **select**: add a `trigger` property to the `wcsFilterChange` event detail to indicate what caused the filter change (user input or option selection)
 - **doc**: 
   - clarify the getting started documentation for Angular regarding the assets folder and paths to static assets
   - add ux guidance for asynchronous options in wcs-select, to indicate to use info `wcs-select-option` during loading